Rename forwarder module file
